Refactors env_name & env_value to re-use extracted reserved_chars_checker type
